+++ projects/amdsmi/amdsmi_cli/amdsmi_parser.py
@@ -366,11 +366,8 @@ class AMDSMIParser(argparse.ArgumentParser):
                 sys.argv[1], string_value, outputformat
             )
 
-    def _is_command_supported(self, user_input, acceptable_values, command_name):
-        if acceptable_values == "N/A":
-            outputformat = self.helpers.get_output_format()
-            raise amdsmi_cli_exceptions.AmdSmiPermissionDeniedException(command_name, outputformat)
-        elif str(user_input).upper() not in acceptable_values:
+    def _is_command_supported(self, user_input, acceptable_values):
+        if str(user_input).upper() not in acceptable_values:
             print(f"Valid inputs are {acceptable_values}")
             raise amdsmi_cli_exceptions.AmdSmiInvalidParameterValueException(
                 sys.argv[1], str(user_input).upper(), self.helpers.get_output_format()
@@ -2522,9 +2519,7 @@ class AMDSMIParser(argparse.ArgumentParser):
                     "--accelerator-partition",
                     action="store",
                     choices=accelerator_set_choices,
-                    type=lambda value: self._is_command_supported(
-                        value, accelerator_set_choices, "--compute-partition"
-                    ),
+                    type=lambda value: self._is_command_supported(value, accelerator_set_choices),
                     required=False,
                     help=set_compute_partition_help,
                     metavar=("TYPE/INDEX"),